add options "viewport" for set_alpha_for_material, so that fade_in can change action in solid viewport shading.
add options "viewport" for set_alpha_for_material, so that fade_in can change action in solid viewport shading.

The default for "viewport" is "material": so that it affects material and render viewport,
if "viewport" is set to "solid", then it will affect solid viewport shading scenarios. This will also change the action for Bobject's appear and disappear actions.